requirements.txt
streamlit==1.39.0
groq>=0.12.0
sentence-transformers==2.7.0
faiss-cpu==1.12.0
pypdf==4.3.1
langdetect==1.0.9
numpy>=2.1.0
pandas==2.2.3
tqdm==4.66.4
streamlit-mic-recorder==0.0.8
gTTS==2.5.4
SQLAlchemy==2.0.36
plotly==5.24.1
httpx==0.27.2
requests>=2.31.0
beautifulsoup4==4.12.3

